Change responsibility for Add Animal - it's only saving new dogs and coming back to previous activity. Change responsibility for Main - it's only showing records and data and it is not able to save data. Removed unnecessary layout Add animal (we are useing firebase test layout.
