+++ app/src/main/java/com/example/android/psiak/data/network/FirebaseRepository.java
@@ -1,7 +1,7 @@
 package com.example.android.psiak.data.network;
 
 import com.example.android.psiak.model.DogFirebase;
-import com.example.android.psiak.ui.addAnimal.FirebaseDataListener;
+import com.example.android.psiak.ui.main.FirebaseDataListener;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
