+++ app/src/main/java/com/example/android/psiak/data/network/Repository.java
@@ -1,6 +1,6 @@
 package com.example.android.psiak.data.network;
 
-import com.example.android.psiak.ui.addAnimal.FirebaseDataListener;
+import com.example.android.psiak.ui.main.FirebaseDataListener;
 import com.example.android.psiak.model.DogFirebase;
 
 import java.util.ArrayList;
